phpstan fixes
