+++ src/Firebase/Factory.php
@@ -684,6 +684,6 @@ class Factory
         $identityApiClient = new IdentityPlatform\ApiClient($httpClient, $this->projectId, $this->tenantId);
 
 
-        return new IdentityPlatform($identityApiClient, $httpClient);
+        return new IdentityPlatform($identityApiClient);
     }
 }
