+++ src/Firebase/IdentityPlatform/InboundSamlConfig.php
@@ -2,6 +2,7 @@
 namespace Kreait\Firebase\IdentityPlatform;
 
 use Kreait\Firebase\Exception\InvalidArgumentException;
+use Kreait\Firebase\IdentityPlatform\ParsesName;
 
 class InboundSamlConfig
 {
@@ -15,11 +16,11 @@ class InboundSamlConfig
 
     public const FIELDS = ['name', 'idpConfig', 'spConfig', 'displayName', 'enabled'];
 
-    private function __construct()
+    final private function __construct()
     {
     }
 
-    public static function new(): self
+    public static function new(): static
     {
         return new static();
     }
@@ -29,7 +30,7 @@ class InboundSamlConfig
      *
      * @throws InvalidArgumentException when invalid properties have been provided
      */
-    public static function withProperties(array $properties): self
+    public static function withProperties(array $properties): static
     {
         $instance = new static();
